requirements.txt
asgiref==3.2.8
Django==3.0.6
django-advanced-filters==1.2.0
django-braces==1.13.0
psycopg2==2.8.5
pytz==2020.1
simplejson==3.17.0
sqlparse==0.3.1
